pyproject.toml
[project]
name = "custom-clip-finder"
version = "2.0.0"
description = "AI-powered viral clip extraction system using 6-AI ensemble"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Jervin Quisada", email = "jervin@quio.agency" }
]

dependencies = [
    # AI APIs
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    
    # Vector Store
    "chromadb>=0.5.0",
    
    # Video Processing
    "ffmpeg-python>=0.2.0",
    
    # Audio Transcription
    "assemblyai>=0.30.0",
    
    # Data Processing
    "pydantic>=2.0.0",
    
    # CLI & Utils
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    
    # HTTP
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[project.scripts]
clip-finder = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["brain", "pipeline", "models", "prompts", "utils"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

